Update favorites.js
Edits to the logic for showing a user's favorites. You will need template support for these changes. Instead of sending all of the user info to the view, it would be better to implement virtuals - a set of public attributes for an object that you can define. Otherwise you'll send the entire user object back to the client, which may be a security risk. So when viewing your own favorites, your routes may send one set of attributes, and when viewing another user's favorites, you may send another set of attributes. (In this app, you'll probably send the same thing for your own, and someone else's favorites). Example of virtuals in the favorite color app:   https://github.com/minneapolis-edu/simple-auth-fav-color/tree/virtuals